Started implement edit and delete pages for categories.

+++ src/java/ro/satrapu/homebudget/services/persistence/PersistenceServiceBean.java
@@ -25,23 +25,39 @@ public class PersistenceServiceBean
 
     @Override
     public <T extends Serializable> T persist(T entity) {
+        if (entity == null) {
+            throw new RuntimeException("Cannot persist null entity");
+        }
+
         entityManager.persist(entity);
         return entity;
     }
 
     @Override
     public <T extends Serializable> void remove(T entity) {
+        if (entity == null) {
+            throw new RuntimeException("Cannot remove null entity");
+        }
+
         T mergedEntity = entityManager.merge(entity);
         entityManager.remove(mergedEntity);
     }
 
     @Override
     public <T extends Serializable> T merge(T entity) {
+        if (entity == null) {
+            throw new RuntimeException("Cannot merge null entity");
+        }
+
         return entityManager.merge(entity);
     }
 
     @Override
     public <T extends Serializable> List<T> listAll(Class<T> entityClass) {
+        if (entityClass == null) {
+            throw new RuntimeException("Cannot query for entities by using null as entity class");
+        }
+
         CriteriaBuilder builder = entityManager.getCriteriaBuilder();
         CriteriaQuery<T> criteria = builder.createQuery(entityClass);
         Root<T> root = criteria.from(entityClass);
@@ -53,6 +69,10 @@ public class PersistenceServiceBean
 
     @Override
     public <T extends Serializable> EntityList<T> list(Class<T> entityClass, int firstResult, int maxResults) {
+        if (entityClass == null) {
+            throw new RuntimeException("Cannot query for entities by using null as entity class");
+        }
+
         CriteriaBuilder builder = entityManager.getCriteriaBuilder();
         CriteriaQuery<T> criteria = builder.createQuery(entityClass);
         Root<T> root = criteria.from(entityClass);
@@ -72,6 +92,14 @@ public class PersistenceServiceBean
 
     @Override
     public <T extends Serializable> T find(Class<T> entityClass, Serializable entityId) {
+        if (entityClass == null) {
+            throw new RuntimeException("Cannot find entity by using null as entity class");
+        }
+
+        if (entityId == null) {
+            throw new RuntimeException("Cannot find entity by using null as entity id");
+        }
+
         return entityManager.find(entityClass, entityId);
     }
 }
